requirements.txt
google-api-python-client~=2.160.0
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
numpy~=2.1.3
moviepy~=2.1.2
pandas~=2.2.3
tqdm~=4.67.1
matplotlib~=3.10.0
google-cloud-storage
gspread~=6.1.4
python-dateutil~=2.9.0.post0
pyairtable~=3.0.2
oauth2client~=4.1.3
pytz~=2025.1
ffmpeg-python
torch~=2.6.0